requirements.txt
openai>=1.0.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
chromadb>=0.5.0
sentence-transformers>=2.7.0
einops>=0.8.0
dateparser>=1.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
gradio>=4.0.0
numpy>=1.26.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
huggingface_hub>=0.20.0
duckduckgo-search>=6.0.0

# MLX libraries (Apple Silicon only - Mac required for voice/local LLM)
mlx-lm>=0.30.0
mlx-audio>=0.1.0
mlx-whisper>=0.3.0
kokoro-onnx>=0.4.0
